requirements.txt
gunicorn==23.0.0
bleach==6.2.0
blinker==1.9.0
click==8.1.7
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
python-dotenv==1.0.1
Unidecode==1.3.8
webencodings==0.5.1
Werkzeug==3.1.3
